#7 moved page DOM classes to subpackage
